fix: refactor Tus parallel logic (#795)
* fix: misc for tus-sender parallel improvements wip
* fix: chunk-sender respect parallel value while including in-progress
* finish all tus e2e tests
* fix: finalize tus parallel refactor
* fix: add logs to bundle size report
* fix: incorrect calc of diff in bundle report

fixes #777

+++ scripts/utils.mjs
@@ -86,7 +86,7 @@ const parseFileSize = (sizeStr) => {
     const match = sizeStr.match(FILE_SIZE_PARSE_RGX);
     if (match) {
         const [_, val, unit] = match;
-        return parseInt(val) * FILESIZE_UNITS[unit.toLowerCase()];
+        return parseFloat(val) * FILESIZE_UNITS[unit.toLowerCase()];
     }
 
     return NaN;
